+++ packages/iov-ethereum/src/ethereumconnector.ts
@@ -6,18 +6,17 @@ import { EthereumConnection, EthereumConnectionOptions } from "./ethereumconnect
 /**
  * A helper to connect to a ethereum-based chain at a given url
  *
- * @param options An EthereumConnectionOptions object. If undefined, all possible options are default.
+ * @param options An EthereumConnectionOptions object. If {}, all possible options are set to their default.
  */
 export function ethereumConnector(
   url: string,
-  options: EthereumConnectionOptions | undefined,
+  options: EthereumConnectionOptions,
   expectedChainId?: ChainId,
 ): ChainConnector {
-  const usedOptions: EthereumConnectionOptions = options || {};
   return {
-    client: async () => EthereumConnection.establish(url, usedOptions),
+    client: async () => EthereumConnection.establish(url, options),
     codec: new EthereumCodec({
-      erc20Tokens: usedOptions.erc20Tokens,
+      erc20Tokens: options.erc20Tokens,
     }),
     expectedChainId: expectedChainId,
   };