+++ CHANGELOG.md
@@ -25,6 +25,8 @@ Breaking changes
 - @iov/bcp: Rename `BcpTicker` to `Token`.
 - @iov/bcp: Rename `BcpConnection.getTicker` to `getToken` and
   `BcpConnection.getAllTickers` to `.getAllTokens`.
+- @iov/ethereum: Options argument in `ethereumConnector` is now required. Can be
+  `{}`.
 
 ## 0.13.3